--- a/pkg/policy/mapstate.go
+++ b/pkg/policy/mapstate.go
@@ -16,6 +16,7 @@ package policy
 
 import (
 	"github.com/cilium/cilium/pkg/identity"
+	"github.com/cilium/cilium/pkg/identity/cache"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy/trafficdirection"
 )
@@ -86,5 +87,19 @@ func (keys MapState) DetermineAllowLocalhost(l4Policy *L4Policy) {
 
 	if option.Config.AlwaysAllowLocalhost() || (l4Policy != nil && l4Policy.HasRedirect()) {
 		keys[localHostKey] = MapStateEntry{}
+	}
+}
+
+// AllowAllIdentities translates all identities in identityCache to their
+// corresponding Key in the specified direction (ingress, egress) which allows
+// all at L3.
+func (keys MapState) AllowAllIdentities(identityCache cache.IdentityCache, direction trafficdirection.TrafficDirection) {
+	// Allow all identities
+	for identity := range identityCache {
+		keyToAdd := Key{
+			Identity:         identity.Uint32(),
+			TrafficDirection: direction.Uint8(),
+		}
+		keys[keyToAdd] = MapStateEntry{}
 	}
 }