--- a/pkg/policy/policy.go
+++ b/pkg/policy/policy.go
@@ -68,6 +68,11 @@ type SearchContext struct {
 	From    labels.LabelArray
 	To      labels.LabelArray
 	DPorts  []*models.Port
+	// rulesSelect specifies whether or not to check whether a rule which is
+	// being analyzed using this SearchContext matches either From or To.
+	// This is used to avoid using EndpointSelector.Matches() if possible,
+	// since it is costly in terms of performance.
+	rulesSelect bool
 }
 
 func (s *SearchContext) String() string {