--- a/pkg/policy/repository.go
+++ b/pkg/policy/repository.go
@@ -24,6 +24,7 @@ import (
 	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy/api"
+	"github.com/cilium/cilium/pkg/policy/trafficdirection"
 )
 
 // Repository is a list of policy rules which in combination form the security
@@ -605,6 +606,29 @@ func (p *Repository) ResolvePolicy(id uint16, labels labels.LabelArray, policyOw
 
 		calculatedPolicy.CIDRPolicy.Ingress = newCIDRIngressPolicy.Ingress
 		calculatedPolicy.L4Policy.Ingress = newL4IngressPolicy.Ingress
+
+		for identity, labels := range identityCache {
+			ingressCtx.From = labels
+			egressCtx.To = labels
+
+			ingressAccess := matchingRules.canReachIngressRLocked(&ingressCtx)
+			if ingressAccess == api.Allowed {
+				keyToAdd := Key{
+					Identity:         identity.Uint32(),
+					TrafficDirection: trafficdirection.Ingress.Uint8(),
+				}
+				calculatedPolicy.PolicyMapState[keyToAdd] = MapStateEntry{}
+			}
+		}
+	} else {
+		// Allow all identities
+		for identity := range identityCache {
+			keyToAdd := Key{
+				Identity:         identity.Uint32(),
+				TrafficDirection: trafficdirection.Ingress.Uint8(),
+			}
+			calculatedPolicy.PolicyMapState[keyToAdd] = MapStateEntry{}
+		}
 	}
 
 	if egressEnabled {
@@ -620,6 +644,28 @@ func (p *Repository) ResolvePolicy(id uint16, labels labels.LabelArray, policyOw
 
 		calculatedPolicy.CIDRPolicy.Egress = newCIDREgressPolicy.Egress
 		calculatedPolicy.L4Policy.Egress = newL4EgressPolicy.Egress
+
+		for identity, labels := range identityCache {
+			egressCtx.To = labels
+
+			egressAccess := matchingRules.canReachEgressRLocked(&egressCtx)
+			if egressAccess == api.Allowed {
+				keyToAdd := Key{
+					Identity:         identity.Uint32(),
+					TrafficDirection: trafficdirection.Egress.Uint8(),
+				}
+				calculatedPolicy.PolicyMapState[keyToAdd] = MapStateEntry{}
+			}
+		}
+	} else {
+		// Allow all identities
+		for identity := range identityCache {
+			keyToAdd := Key{
+				Identity:         identity.Uint32(),
+				TrafficDirection: trafficdirection.Egress.Uint8(),
+			}
+			calculatedPolicy.PolicyMapState[keyToAdd] = MapStateEntry{}
+		}
 	}
 
 	calculatedPolicy.computeDesiredL4PolicyMapEntries(identityCache)