Based on ~~#6275~~ .  I'm waiting for that to get merged before merging this. I rebased on top of it to avoid conflicts. 

This PR is part of the overall work to improve the performance / scalability of policy calculation within Cilium. In order to make testing scalability and performance of policy easier, it behooves us to factor out policy generation for a given set of labels (i.e., the security-relevant labels for an endpoint which comprise its security identity) into in `pkg/policy`. 

Add a new type which encompasses the policy at L3, L4, and L7 that is generated based off of the provided set of labels. This currently wraps the pre-existing types that were distinct fields within an `Endpoint`, `L3Policy`, `{Desired,Realized}L4Policy`, and `{desired,realized}PolicyMapState`. This is done so that in the future, we can hide how policy is plumbed from the endpoint structure itself, allowing for an iterative refactor of the internals of this new type to improve performance of policy regeneration.

This PR improves policy calculation performance as well. When the initial pass is performed over the rules to determine whether ingress and egress policy enforcement apply for the provided set of labels, the list of rules which select the endpoint itself is also returned. By returning this subset of the rules in the policy repository, when we pass over the rules to generate policy for each type of policy object generated (`CIDRPolicy`, `L4Policy`, `MapState`), we no longer have to check whether the rules being analyzed select the given set of labels. Caching the rules that match locally improves performance, because performing matching operations against `EndpointSelector` is costly; this reduces the number of times matching is performed against the `EndpointSelector` to be done only once during policy calculation. It also means that policy generation is only performed against the subset of rules which select the set of labels. In the case in which only one rule selects the set of labels out of thousands of rules, not only is only that rule considered when the policy rules are translated, we no longer perform matching operations against each and every one of those rules which don't even select the endpoint multiple times.

We track whether or not to check whether the EndpointSelector in the rule matches the set of provided labels by providing a new field in the `SearchContext`, called `rulesSelect`. If set, the code does not check whether the `EndpointSelector` selects the given set of labels, allowing matching to be skipped.

I ran benchmarks against `regeneratePolicy`, and found that this leads to around 54 % decrease in amount of time taken to regenerate policy in the case where no rules select the endpoint for which policy is being regenerated:

Previous:
```
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=false ipv4= ipv6= k8sPodName=/ policyChanged=true policyRegenerationTime=562.150107ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyChanged=false policyRegenerationTime=451.340492ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyChanged=false policyRegenerationTime=696.477095ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyChanged=false policyRegenerationTime=661.700047ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyChanged=false policyRegenerationTime=567.049182ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyChanged=false policyRegenerationTime=465.258572ms subsys=endpoint
PASS: policy_test.go:215: EndpointSuite.BenchmarkRegeneratePolicyRules	       5	 568447601 ns/op
```

With changes in this PR:

```
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=false ipv4= ipv6= k8sPodName=/ policyRegenerationTime=197.632603ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=205.555571ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=222.051934ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=441.657516ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=203.362918ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=206.866816ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=415.307384ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=206.394149ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=227.010221ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=317.161773ms subsys=endpoint
level=info msg="Completed endpoint policy recalculation" containerID= datapathPolicyRevision=0 desiredPolicyRevision=1 endpointID=256 forcedRegeneration=true ipv4= ipv6= k8sPodName=/ policyRegenerationTime=201.965828ms subsys=endpoint
PASS: policy_test.go:215: EndpointSuite.BenchmarkRegeneratePolicyRules	      10	 264943078 ns/op
OK: 1 passed
```

Things this PR does not do:

* Change the Endpoint structure itself. Ideally we can package this new `Policy` type as a field within the endpoint. I'd like to keep that to a future PR to avoid having too much going on in this PR. 
* Unify policy generation. The same structures as before (`L4Policy`, `CIDRPolicy`, BPF PolicyMap state) are kept the same as before. Changing of these structures will be done by future PRs. 
* Add unit tests for checking the state of a `Policy` (yet). I didn't want to add a ton of tests for something which would be changed. FWIW, the results of it are only used in `Endpoint`, and the CI has passed, which was the main test for `regeneratePolicy` before. I can add these in the future / as follow-up commits if people agree on the general change I'm making here.

Signed-off by: Ian Vernon <ian@cilium.io>

```release-note
Increase performance of policy computation when rules don't select labels by approximately 50 %
```

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cilium/cilium/6270)
<!-- Reviewable:end -->