@ianvernon 2018-11-27
Build failed due to downloading of VM being slow again :(