@tgraf 2018-12-04
#6275 was merged