@ianvernon 2018-12-04
@joestringer thanks for the feedback re: the removal / moving of code. I'll be sure to be more mindful of that in the future!